requirements.txt
slack-sdk>=3.33.1,<4
slack-bolt==1.21.2
python-dotenv==1.0.0
gunicorn==23.0.0
pandas>=2.0.0
flask>=3.0.0
